LIBIIIF-129. Initial implementation of iiif-static-files
Created "iiif-static-files" repository, which is used to generate an
"iiif-static-files" Docker image for serving the IIIF home page
and Mirador viewer static files.

The Docker image is based on the "umd-web-static-files" Docker image,
from which it gets the "robot.txt", "sitemap.xml" and "favicon.ico"
files.

Several versions of the Mirador viewer
(from https://github.com/umd-lib/mirador-static) are included, with the
".git" subdirectories removed to avoid adding unnecessary bloat to the
Docker image.

https://issues.umd.edu/browse/LIBIIIF-129