LIBIIIF-128. Updated Nginx configuration to proxy to legacy IIIF servers
Replaced the default Nginx configuration from "umd-web-static-files"
with a configuration that enables the "/images" path to be proxied
to a legacy IIIF server.

Because Nginx does not have a mechanism for using environment variables
in its configuration files (see https://serverfault.com/a/755541), and
the only workaround was to use "envsubst", which didn't seem like an
easy fit for what we're trying to do, a Kubernetes "ExternalName"
service (in the "umd-lib/k8s-iiif" repository) is used to handle the
redirect to the correct server.

There did not appear to be any way to merge a modified Nginx
configuration with the Nginx configuration from "umd-web-static-files",
so the Nginx configuration from "umd-web-static-files" is completely
replaced with the repository's own copy of the
"docker_config/nginx/conf.d/default.conf" file.

https://issues.umd.edu/browse/LIBIIIF-128